Rename to `LazyPtr` and move to the `lazy` module
