+++ src/netbsd.rs
@@ -1,5 +1,5 @@
 //! Implementation for NetBSD
-use crate::{cached_ptr::CachedPtr, util_libc::sys_fill_exact, Error};
+use crate::{lazy::LazyPtr, util_libc::sys_fill_exact, Error};
 use core::{ffi::c_void, mem::MaybeUninit, ptr};
 
 fn kern_arnd(buf: &mut [MaybeUninit<u8>]) -> libc::ssize_t {
@@ -25,7 +25,7 @@ fn kern_arnd(buf: &mut [MaybeUninit<u8>]) -> libc::ssize_t {
 type GetRandomFn = unsafe extern "C" fn(*mut u8, libc::size_t, libc::c_uint) -> libc::ssize_t;
 
 // getrandom(2) was introduced in NetBSD 10.0
-static GETRANDOM: CachedPtr = CachedPtr::new();
+static GETRANDOM: LazyPtr = LazyPtr::new();
 
 fn dlsym_getrandom() -> *mut c_void {
     static NAME: &[u8] = b"getrandom\0";
@@ -34,7 +34,8 @@ fn dlsym_getrandom() -> *mut c_void {
 }
 
 pub fn getrandom_inner(dest: &mut [MaybeUninit<u8>]) -> Result<(), Error> {
-    if let Some(fptr) = GETRANDOM.get(dlsym_getrandom) {
+    let fptr = GETRANDOM.unsync_init(dlsym_getrandom);
+    if !fptr.is_null() {
         let func: GetRandomFn = unsafe { core::mem::transmute(fptr) };
         return sys_fill_exact(dest, |buf| unsafe {
             func(buf.as_mut_ptr() as *mut u8, buf.len(), 0)
